formatFileSize() must use only the language, not the locale.
formatFileSize() use Number.toLocaleString(LOCALE) and passes
getCanonicalLocale() to it. This is wrong, as at least for recent FF
and Chromium browsers toLocaleString() only accepts the language tag,
see for example

https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString

and the demo on that site.

However, getCanonicalLocale() finally return $('html').data('locale'),
replacing _ by _. This results for example in LOCALE=de-DE.UTF-8 and
leads to a

Error: invalid language tag: "de-DE.UTF-8"

The bug is only triggered when the first digit of the fractional part
of the size is != 0, which probably is the reason why the bug is not
triggered so often.

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>